# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Documentation from DASA Challenge API",
    "description": "This docs is about exams and laboratories",
    "version": "1.0.0",
    "contact": {
      "email": "joaobispo2077@gmail.com"
    }
  },
  "paths": {
    "/laboratories/{id}/exams": {
      "post": {
        "tags": ["Laboratories"],
        "summary": "Add relationship between laboratory and exam",
        "description": "Add relationship between laboratory and exam",
        "operationId": "addExamToLaboratory",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Laboratory id"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "examId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/laboratories": {
      "post": {
        "tags": ["Laboratories"],
        "summary": "Create a laboratory",
        "description": "Create a new laboratory",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/LaboratoryInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      },
      "get": {
        "tags": ["Laboratories"],
        "summary": "Get all laboratories ",
        "description": "Get all laboratories active",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "array": {
                    "items": {
                      "$ref": "#/components/LaboratoryOutput"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Laboratories"],
        "summary": "Delete one laboratory by id",
        "description": "Delete one laboratory by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Laboratory id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "patch": {
        "tags": ["Laboratories"],
        "summary": "Update one laboratory by id",
        "description": "Update one laboratory by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Laboratory id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/LaboratoryUpdateInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/LaboratoryOutput"
                }
              }
            }
          }
        }
      }
    },
    "/exams": {
      "post": {
        "tags": ["Exams"],
        "summary": "Create an exam",
        "description": "Create a new exam",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/ExamInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      },
      "get": {
        "tags": ["Exams"],
        "summary": "Get all exams ",
        "description": "Get all exams active",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "array": {
                    "items": {
                      "$ref": "#/components/ExamOutput"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Exams"],
        "summary": "Delete one exam by id",
        "description": "Delete one exam by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Exam id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "patch": {
        "tags": ["Exams"],
        "summary": "Update one exam by id",
        "description": "Update one exam by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Exam id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/ExamUpdateInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/ExamOutput"
                }
              }
            }
          }
        }
      }
    },
    "/ready": {
      "get": {
        "tags": ["System"],
        "summary": "Returns message if system is ready",
        "description": "Returns is ready message",
        "operationId": "getReady",
        "responses": {
          "200": {
            "description": "Successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/version": {
      "get": {
        "tags": ["System"],
        "summary": "Returns version of the system",
        "description": "Returns version of the system",
        "operationId": "getVersion",
        "responses": {
          "200": {
            "description": "Successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string"
                    },
                    "author": {
                      "type": "string"
                    },
                    "version": {
                      "type": "string"
                    },
                    "description": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "LaboratoryInput": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "address": {
          "type": "string"
        }
      },
      "example": {
        "name": "Laboratório Paulista",
        "address": "Rua Paulista, 100",
        "status": "ativo"
      }
    },
    "LaboratoryUpdateInput": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "address": {
          "type": "string"
        },
        "status": {
          "type": "string"
        }
      },
      "example": {
        "name": "Laboratório Paulista",
        "address": "Rua Paulista, 300",
        "status": "inativo"
      }
    },
    "LaboratoryOutput": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "address": {
          "type": "string"
        },
        "status": {
          "type": "string"
        }
      },
      "example": {
        "id": "9f0838a3-14f0-4677-9ffe-2332baed5dfa",
        "name": "Laboratório Paulista",
        "address": "Rua Paulista, 100",
        "status": "ativo"
      }
    },
    "ExamInput": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "type": {
          "type": "string"
        }
      },
      "example": {
        "id": "9f0838a3-14f0-4677-9ffe-2332baed5dfa",
        "name": "Exame de sangue",
        "type": "analise clinica"
      }
    },
    "ExamUpdateInput": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "type": {
          "type": "string"
        },
        "status": {
          "type": "string"
        }
      },
      "example": {
        "name": "Exame de sangue",
        "type": "analise clinica",
        "status": "ativo"
      }
    },
    "ExamOutput": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "type": {
          "type": "string"
        },
        "status": {
          "type": "string"
        }
      },
      "example": {
        "id": "32c1bbc9-d173-42e7-8bae-e992c2a7b945",
        "name": "Raio X",
        "type": "imagem",
        "status": "ativo"
      }
    },
    "LaboratoyExamCreateOuput": {
      "id": {
        "type": "string"
      },
      "laboratoryId": {
        "type": "string"
      },
      "examId": {
        "type": "string"
      },
      "example": {
        "id": "b226113d-ff7f-41e2-ab0d-4919efeb22df",
        "laboratoryId": "c0013d52-f64f-4353-a578-0ddab0872f27",
        "examId": "6d43992a-aa15-4e0a-8ab9-e6dbd7c06315"
      }
    }
  }
}